[refactor] move driverResolver/driverWrapper -> driver.js
