--- packages/cli-poa/src/driverResolver/driver.js
@@ -1,9 +1,0 @@
-import DriverWrapper from './driverWrapper.js';
-
-export default class Driver {
-  constructor(sessionId, executorUrl) {
-    this.sessionId = sessionId;
-    this.executorUrl = executorUrl;
-    this.helper = new DriverWrapper(this.sessionId, this.executorUrl);
-  }
-}
