+++ packages/cli-poa/src/providers/automateProvider.js
@@ -21,7 +21,7 @@ export default class AutomateProvider extends GenericProvider {
 
   async browserstackExecutor(action, args) {
     let options = args ? { action, arguments: args } : { action };
-    let res = await this.driver.helper.executeScript({script:`browserstack_executor: ${JSON.stringify(options)}`, args: []}); 
+    let res = await this.driver.executeScript({script:`browserstack_executor: ${JSON.stringify(options)}`, args: []}); 
     return res
   }
